refactor: move fieldError from server -> shared
this is because that the shared folder should not inherit from either the client or server folder
