+++ src/shared/model/ClientState.ts
@@ -1,9 +1,12 @@
-import { FieldError } from '@/server/routes/changePassword';
 import { Consent } from '@/shared/model/Consent';
 import { NewsLetter } from '@/shared/model/Newsletter';
 import { GeoLocation } from '@/shared/model/Geolocation';
 import { Participations } from '@guardian/ab-core';
 
+export interface FieldError {
+  field: string;
+  message: string;
+}
 interface ABTesting {
   mvtId?: number;
   participations?: Participations;
